vm-runtime: remove deps on starcoin-config
