+++ vm/compiler/Cargo.toml
@@ -8,13 +8,11 @@ version = "1.12.1"
 
 [dependencies]
 anyhow = "1.0.41"
-itertools = "0.10.3"
 move-binary-format = {git = "https://github.com/starcoinorg/move", rev = "9d1e76b770d630ad646a540b325f88ef986af3a3"}
 move-command-line-common = {git = "https://github.com/starcoinorg/move", rev = "9d1e76b770d630ad646a540b325f88ef986af3a3"}
 move-compiler = {package = "move-compiler", git = "https://github.com/starcoinorg/move", rev = "9d1e76b770d630ad646a540b325f88ef986af3a3"}
 once_cell = "1.13.0"
 petgraph = "0.5.1"
-rayon = "1.5.1"
 regex = {version = "1.6.0", default-features = false, features = [
   "std",
   "perf",
