+++ vm/stdlib/Cargo.toml
@@ -10,7 +10,7 @@ version = "1.12.1"
 anyhow = "1.0.41"
 bcs-ext = {package = "bcs-ext", path = "../../commons/bcs_ext"}
 clap = {version = "3", features = ["derive"]}
-fs_extra = "1.2.0"
+
 include_dir = "0.6.2"
 itertools = "0.10.3"
 log = "0.4.16"
@@ -27,6 +27,7 @@ starcoin-move-compiler = {path = "../../vm/compiler"}
 starcoin-vm-types = {path = "../types"}
 tempfile = "3.2.0"
 walkdir = "2.3"
+fs_extra = "1.2.0"
 
 [dev-dependencies]
 datatest-stable = "0.1.1"
