+++ vm/types/Cargo.toml
@@ -17,7 +17,6 @@ num_enum = "0.5.7"
 once_cell = "1.13.0"
 rand = "0.8.5"
 serde = {version = "1.0.130", default-features = false}
-serde-helpers = {path = "../../commons/serde-helpers"}
 serde_bytes = "0.11.5"
 serde_json = {version = "1.0", features = ["arbitrary_precision"]}
