+++ vm/types/src/lib.rs
@@ -6,55 +6,8 @@ mod language_storage_ext;
 
 pub mod account_address;
 
-pub mod gas_schedule {
-    pub use move_core_types::gas_schedule::*;
-    pub use move_vm_types::gas_schedule::{
-        calculate_intrinsic_gas, new_from_instructions, zero_cost_schedule, GasStatus,
-    };
-    #[allow(non_camel_case_types)]
-    #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq, PartialOrd, Ord)]
-    #[repr(u8)]
-    pub enum NativeCostIndex {
-        SHA2_256 = 0,
-        SHA3_256 = 1,
-        ED25519_VERIFY = 2,
-        ED25519_THRESHOLD_VERIFY = 3,
-        BCS_TO_BYTES = 4,
-        LENGTH = 5,
-        EMPTY = 6,
-        BORROW = 7,
-        BORROW_MUT = 8,
-        PUSH_BACK = 9,
-        POP_BACK = 10,
-        DESTROY_EMPTY = 11,
-        SWAP = 12,
-        ED25519_VALIDATE_KEY = 13,
-        SIGNER_BORROW = 14,
-        CREATE_SIGNER = 15,
-        DESTROY_SIGNER = 16,
-        EMIT_EVENT = 17,
-        BCS_TO_ADDRESS = 18,
-        TOKEN_NAME_OF = 19,
-        KECCAK_256 = 20,
-        RIPEMD160 = 21,
-        ECRECOVER = 22,
-        U256_FROM_BYTES = 23,
-        U256_ADD = 24,
-        U256_SUB = 25,
-        U256_MUL = 26,
-        U256_DIV = 27,
-        U256_REM = 28,
-        U256_POW = 29,
-        VEC_APPEND = 30,
-        VEC_REMOVE = 31,
-        VEC_REVERSE = 32,
-    }
+pub mod gas_schedule;
 
-    impl NativeCostIndex {
-        //note: should change this value when add new native function.
-        pub const NUMBER_OF_NATIVE_FUNCTIONS: usize = 33;
-    }
-}
 pub mod location {
     pub use move_ir_types::location::Loc;
 }
