+++ vm/vm-runtime/Cargo.toml
@@ -18,18 +18,16 @@ move-vm-runtime = {git = "https://github.com/starcoinorg/move", rev = "9d1e76b77
 num_enum = "0.5.7"
 rand = "0.8.5"
 rand_core = {version = "0.6.3", default-features = false}
-starcoin-config = {path = "../../config"}
 starcoin-logger = {path = "../../commons/logger"}
 starcoin-metrics = {path = "../../commons/metrics"}
-starcoin-move-compiler = {path = "../compiler"}
 starcoin-natives = {path = "../natives"}
 starcoin-types = {path = "../../types"}
 starcoin-vm-types = {path = "../types"}
-stdlib = {package = "stdlib", path = "../stdlib"}
+
 tracing = "0.1.34"
 
 [dev-dependencies]
-
+stdlib = {package = "stdlib", path = "../stdlib"}
 [features]
 default = []
 testing = ["move-stdlib/testing", "starcoin-natives/testing"]
