+++ vm/vm-runtime/src/starcoin_vm.rs
@@ -14,7 +14,6 @@ use move_vm_runtime::move_vm::MoveVM;
 use move_vm_runtime::move_vm_adapter::{PublishModuleBundleOption, SessionAdapter};
 use move_vm_runtime::session::Session;
 use once_cell::sync::Lazy;
-use starcoin_config::G_LATEST_GAS_SCHEDULE;
 use starcoin_logger::prelude::*;
 use starcoin_types::account_config::config_change::ConfigChangeEvent;
 use starcoin_types::account_config::{
@@ -39,6 +38,7 @@ use starcoin_vm_types::account_config::{
 use starcoin_vm_types::contract_event::ContractEvent;
 use starcoin_vm_types::file_format::{CompiledModule, CompiledScript};
 use starcoin_vm_types::gas_schedule::NativeCostIndex;
+use starcoin_vm_types::gas_schedule::G_LATEST_GAS_SCHEDULE;
 use starcoin_vm_types::gas_schedule::{zero_cost_schedule, GasConstants, GasCost, GasStatus};
 use starcoin_vm_types::genesis_config::StdlibVersion;
 use starcoin_vm_types::identifier::IdentStr;
